Create .py
It's a Python program that acts as a virtual Magic 8-Ball, offering a distinct fortune in response to any 'Yes' or 'No' question. Each time the program is run, it will randomly select one of the nine possible answers from a predefined list. As a result, every execution of the program will yield a different fortune, adding an element of surprise and unpredictability to the responses.